+++ fabric/src/main/java/com/dudko/blazinghot/multiloader/fabric/MultiRegistriesImpl.java
@@ -7,6 +7,7 @@ import org.jetbrains.annotations.NotNull;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.material.Fluid;
 
 public class MultiRegistriesImpl {
@@ -17,5 +18,10 @@ public class MultiRegistriesImpl {
 	@NotNull
 	public static Supplier<Fluid> getFluidFromRegistry(ResourceLocation resourceLocation) {
 		return () -> BuiltInRegistries.FLUID.get(resourceLocation);
+	}
+
+	@NotNull
+	public static Supplier<Block> getBlockFromRegistry(ResourceLocation resourceLocation) {
+		return () -> BuiltInRegistries.BLOCK.get(resourceLocation);
 	}
 }