+++ forge/src/main/java/com/dudko/blazinghot/multiloader/forge/MultiRegistriesImpl.java
@@ -6,6 +6,7 @@ import org.jetbrains.annotations.NotNull;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -17,5 +18,10 @@ public class MultiRegistriesImpl {
 	@NotNull
 	public static Supplier<Fluid> getFluidFromRegistry(ResourceLocation resourceLocation) {
 		return () -> ForgeRegistries.FLUIDS.getValue(resourceLocation);
+	}
+
+	@NotNull
+	public static Supplier<Block> getBlockFromRegistry(ResourceLocation resourceLocation) {
+		return () -> ForgeRegistries.BLOCKS.getValue(resourceLocation);
 	}
 }